Implements database migrations
