+++ src/main/java/org/sgrewritten/stargate/migration/DataMigration.java
@@ -1,7 +1,9 @@
 package org.sgrewritten.stargate.migration;
 
 import org.bukkit.configuration.ConfigurationSection;
+import org.jetbrains.annotations.NotNull;
 import org.sgrewritten.stargate.container.TwoTuple;
+import org.sgrewritten.stargate.database.SQLDatabaseAPI;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -43,7 +45,7 @@ public abstract class DataMigration {
      * <p>This may alter files such as portal and gate files to perform the appropriate changes necessary for
      * compatibility. This does not modify the config file itself.</p>
      */
-    public abstract void run();
+    public abstract void run(@NotNull SQLDatabaseAPI base);
 
     /**
      * Gets the configuration version this modifier will update the configuration to
